refactor: modularization
